chapter 10 is done

+++ part3/src/data/creature.py
@@ -1,15 +1,19 @@
 import sys
 sys.path.append('/mnt/c/Users/hyoja/OneDrive/문서/GitHub/fastAPI_for_the_first_time/part3/src')
 
-from .init import conn, curs
+from .init import conn, curs, IntegrityError
 from model.creature import Creature  
+from error import Missing, Duplicate
 
-curs.execute("""create table if not exists creature(
-             name text primary key,
-             description text,
-             country text,
-             area text,
-             aka text)""")
+curs.execute(
+    """create table if not exists creature(
+        name text primary key,
+        description text,
+        country text,
+        area text,
+        aka text)"""
+    )
+
 
 def row_to_model(row: tuple) -> Creature:
     name, description, country, area, aka = row  
@@ -21,30 +25,42 @@ def row_to_model(row: tuple) -> Creature:
         aka=aka,
     )
 
+
 def model_to_dict(creature: Creature) -> dict:
     return creature.model_dump() # model_dump : 모델 인스턴스를 dict 타입으로 변경하는 메소드 
+
 
 def get_one(name: str) -> Creature:
     qry = "select * from creature where name=:name"
     params = {"name":name}
     curs.execute(qry, params)
-    return row_to_model(curs.fetchone())
+    row = curs.fetchone()
+    if row:
+        return row_to_model(row)
+    else:
+        raise Missing(msg=f"Creature {name} not found")
+
 
 def get_all() -> list[Creature]:
     qry = "select * from creature"
     curs.execute(qry)
     return [row_to_model(row) for row in curs.fetchall()]
 
+
 def create(creature: Creature) -> Creature:
     qry = "insert into creature values" \
         " (:name, :description, :country, :area, :aka)"
     params = model_to_dict(creature)
-    curs.execute(qry, params)
+    try:
+        curs.execute(qry, params)
+    except IntegrityError:
+        raise Duplicate(msg=f"Creature {creature.name} already exists")
     return get_one(creature.name)
 
-def modify(creature: Creature) -> Creature:
+
+def modify(name: str, creature: Creature) -> Creature:
     qry = """update creature
-            set country:=country,
+            set country=:country,
                 name=:name,
                 description=:description,
                 area=:area,
@@ -52,12 +68,36 @@ def modify(creature: Creature) -> Creature:
             where name=:name_orig"""
     params = model_to_dict(creature)
     params["name_orig"] = creature.name
-    _ = curs.execute(qry, params)
-    return get_one(creature.name)
+    curs.execute(qry, params)
+    if curs.rowcount == 1:
+        return get_one(creature.name)
+    else:
+        raise Missing(msg=f"Creature {creature.name} not found")
 
 
-def delete(creature: Creature) -> bool:
+def replace(name: str, creature: Creature) -> Creature:
+    qry = """update creature
+            set country=:country,
+                name=:name,
+                description=:description,
+                area=:area,
+                aka=:aka
+            where name=:name_orig"""
+    params = model_to_dict(creature)
+    params["name_orig"] = creature.name
+    curs.execute(qry, params)
+    if curs.rowcount == 1:
+        return get_one(creature.name)
+    else:
+        raise Missing(msg=f"Creature {name} not found")
+
+
+def delete(name: str) -> bool:
+    if not name:
+        return False 
     qry = "delete from creature where name = :name"
-    params = {"name": creature.name}
-    res = curs.execute(qry, params)
-    return bool(res)
+    params = {"name": name}
+    curs.execute(qry, params)
+    if curs.rowcount != 1:
+        raise Missing(msg=f"Creature {name} not found")
+    return True
